requirements.txt
aiohttp~=3.13.2
kurigram~=2.2.17
pillow~=12.0.0
psutil~=7.2.1
pymongo~=4.15.5
pytgcrypto~=1.2.11
py-tgcalls~=2.2.8
py-yt-search~=0.5.5
python-dotenv~=1.2.1
yt-dlp[default]>=2026.3.3
ffmpeg
boto3>=1.26.0
botocore>=1.29.0
